Refactor with Layout & Headers
