+++ app/components/SearchBar/SearchBar.tsx
@@ -1,0 +1,30 @@
+"use client";
+
+import { useRouter } from "next/navigation";
+import { useState } from "react";
+
+export default function SearchBar() {
+  const router = useRouter();
+  const [location, setLocation] = useState("");
+
+  return (
+    <div className='text-left py-3 m-auto flex justify-center'>
+      <input
+        type='text'
+        className='rounded text-lg mr-3 w-74 p-2 w-[450px] bg-white'
+        placeholder='Location, Restaurant, or Cuisine'
+        value={location}
+        onChange={(e) => setLocation(e.target.value)}
+      />
+      <button
+        className='rounded bg-red-600 px-9 py-2 text-white'
+        onClick={() => {
+          if (location === "banana") return;
+          router.push("/search");
+        }}
+      >
+        Let's go
+      </button>
+    </div>
+  );
+}
