+++ app/layout.tsx
@@ -1,18 +1,26 @@
-import './globals.css'
+import Navbar from "./components/Navbar/Navbar";
+import "./globals.css";
 
 export default function RootLayout({
   children,
 }: {
-  children: React.ReactNode
+  children: React.ReactNode;
 }) {
   return (
-    <html lang="en">
+    <html lang='en'>
       {/*
         <head /> will contain the components returned by the nearest parent
         head.tsx. Find out more at https://beta.nextjs.org/docs/api-reference/file-conventions/head
       */}
       <head />
-      <body>{children}</body>
+      <body>
+        <main className='bg-gray-100 min-h-screen w-screen text-black'>
+          <main className='max-w-screen-2xl m-auto bg-white'>
+            <Navbar />
+            {children}
+          </main>
+        </main>
+      </body>
     </html>
-  )
+  );
 }
