+++ app/search/head.tsx
@@ -1,7 +1,7 @@
 export default function Head() {
   return (
     <>
-      <title>OpenTables</title>
+      <title>Search | OpenTables</title>
       <meta content='width=device-width, initial-scale=1' name='viewport' />
       <meta name='description' content='Generated by create next app' />
       <link rel='icon' href='/favicon.ico' />
